requirements.txt
# Core web framework
Flask==3.0.3
Werkzeug==3.1.3
Jinja2==3.1.4
click>=8.0.2,<9.0

# AI/LLM dependencies (updated for compatibility)
openai==1.97.1
langchain==0.3.26
pydantic>=2.7.4,<3.0.0  # Matches langchain's requirement
llama-index-core==0.12.52.post1
tiktoken==0.9.0

# JSON/data handling
orjson==3.10.7

# HTTP/API clients
requests==2.31.0
httpx==0.27.0

# Environment management
python-dotenv==1.1.1

sentence-transformers
faiss-cpu
numpy

flask
faiss-cpu
sentence-transformers
numpy
werkzeug
Flask
faiss-cpu
sentence-transformers
numpy
gunicorn
